move Arguments to separate package
